backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart==0.0.6
face-recognition==1.3.0
numpy>=1.26.0
websockets==12.0
Pillow==10.1.0
